Lab-02-Podcast/requirements.txt
agent-framework
agent-framework-orchestrations
agent-framework-github-copilot
python-dotenv
